Fixed Flip Card Bug
There was a bug in the code where the card would only flip if the user clicked or hit enter twice.
The bug is now fixed so that the card will flip as long as the user hits enter or clicks the flip card button.

I also added a temporary text field for the flip card button so it's easier to find.